[#1308] Upgrade ExportLogsReducer to the latest TCA
- ExportLogsReducer has been refactored to @Reducer ExportLogs
